+++ src/components.d.ts
@@ -10,16 +10,18 @@ declare module 'vue' {
     AdminMenu: typeof import('./components/admin/AdminMenu.vue')['default']
     Avatar: typeof import('./components/user/Avatar.vue')['default']
     AvatarInfo: typeof import('./components/user/AvatarInfo.vue')['default']
-    CodeEditor: typeof import('./components/CodeEditor.vue')['default']
-    CodeRun: typeof import('./components/CodeRun.vue')['default']
+    CodeEditor: typeof import('./components/judge/CodeEditor.vue')['default']
+    CodeRun: typeof import('./components/judge/CodeRun.vue')['default']
     ElAside: typeof import('element-plus/es')['ElAside']
     ElAvatar: typeof import('element-plus/es')['ElAvatar']
     ElButton: typeof import('element-plus/es')['ElButton']
     ElCard: typeof import('element-plus/es')['ElCard']
     ElCarousel: typeof import('element-plus/es')['ElCarousel']
     ElCarouselItem: typeof import('element-plus/es')['ElCarouselItem']
+    ElCheckTag: typeof import('element-plus/es')['ElCheckTag']
     ElCol: typeof import('element-plus/es')['ElCol']
     ElContainer: typeof import('element-plus/es')['ElContainer']
+    ElDialog: typeof import('element-plus/es')['ElDialog']
     ElDivider: typeof import('element-plus/es')['ElDivider']
     ElForm: typeof import('element-plus/es')['ElForm']
     ElFormItem: typeof import('element-plus/es')['ElFormItem']
@@ -33,17 +35,19 @@ declare module 'vue' {
     ElPopover: typeof import('element-plus/es')['ElPopover']
     ElRow: typeof import('element-plus/es')['ElRow']
     ElSelect: typeof import('element-plus/es')['ElSelect']
+    ElTable: typeof import('element-plus/es')['ElTable']
+    ElTableColumn: typeof import('element-plus/es')['ElTableColumn']
     ElTag: typeof import('element-plus/es')['ElTag']
     HeaderNav: typeof import('./components/HeaderNav.vue')['default']
-    LanguageSelect: typeof import('./components/LanguageSelect.vue')['default']
-    LogoutButton: typeof import('./components/auth/LogoutButton.vue')['default']
+    LanguageSelect: typeof import('./components/judge/LanguageSelect.vue')['default']
+    LogoutButton: typeof import('./components/user/LogoutButton.vue')['default']
     ProblemDifficultySelect: typeof import('./components/problem/ProblemDifficultySelect.vue')['default']
     ProblemList: typeof import('./components/problem/ProblemList.vue')['default']
     ProblemListItem: typeof import('./components/problem/ProblemListItem.vue')['default']
     Problems: typeof import('./components/problem/Problems.vue')['default']
     ProblemSelect: typeof import('./components/problem/ProblemSelect.vue')['default']
     ProblemTagSelect: typeof import('./components/problem/ProblemTagSelect.vue')['default']
-    RecordInfo: typeof import('./components/RecordInfo.vue')['default']
+    RecordInfo: typeof import('./components/record/RecordInfo.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
     ToUserSettingButton: typeof import('./components/user/ToUserSettingButton.vue')['default']
