+++ src/components/record/RecordInfo.vue
